move data files to domain
