+++ web-app/src/api/legislation/index.ts
@@ -1,0 +1,1 @@
+export * from "./api";
