+++ web-app/src/api/representatives/google.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 import type { Env } from "~app/modules/config";
-import { transformGoogleCivicInfo } from "../representatives.utils";
+import { transformGoogleCivicInfo } from "../../../../domain/representatives/representatives.utils";
 import type { GoogleRepresentativesResponse } from "./google.types";
 
 export const getRepresentatives = async (address: string, env: Env) => {
