--- web-app/src/app/modules/data/legislation/api/index.ts
@@ -1,2 +1,0 @@
-export * from "./api";
-export * from "../../types";
