--- web-app/src/app/modules/data/legislation/index.ts
@@ -1,4 +1,0 @@
-export * from "./api";
-export * from "../filters";
-export * from "./legislation.utils";
-export * from "./legislation.types";
