+++ web-app/src/app/modules/feed-ui/feed-ui.loader.tsx
@@ -1,13 +1,13 @@
 import { json, type LoaderFunction } from "react-router-dom";
 import { getEnv } from "~app/modules/config";
 
-import { getFilteredLegislation } from "~app/modules/data/api";
-import type { FilterParams } from "~app/modules/data/filters";
+import { getFilteredLegislation } from "@windycivi/domain/api";
+import type { FilterParams } from "@windycivi/domain/filters";
 import {
   DEFAULT_FILTERS,
   createFilterParams,
   parseRepLevel,
-} from "~app/modules/data/filters";
+} from "@windycivi/domain/filters";
 import { DEFAULT_GLOBAL_STATE, RouteOption } from "./feed-ui.constants";
 import { type FeedLoaderData } from "./feed-ui.types";
 import { getCookieFromString } from "./feed-ui.utils";
