+++ web-app/src/app/modules/feed-ui/react/Bills.tsx
@@ -1,16 +1,13 @@
 import { FaGlobe } from "react-icons/fa";
-import {
-  getBillUpdateAt,
-  WindyCiviBill,
-  RepLevel,
-  getLastStatus,
-  mapToReadableStatus,
-} from "~app/modules/data/legislation";
+
 import { StyleHack, Tag, classNames } from "~app/modules/design-system";
 import { Carousel } from "~app/modules/design-system/Carousel/Carousel";
 import { RobotSvg } from "~app/modules/design-system/Icons";
-import { FeedProps } from "../feed-ui.types";
 import { LevelFilter } from "./Filters";
+import { FeedProps } from "../feed-ui.types";
+import { getBillUpdateAt, RepLevel } from "@windycivi/domain/filters";
+import { WindyCiviBill } from "@windycivi/domain/types";
+import { getLastStatus, mapToReadableStatus } from "@windycivi/domain/legislation/legislation.utils";
 
 const newBillGlow = {
   filter: "drop-shadow(0px 0px 8px rgb(59, 130, 246))",
