+++ web-app/src/app/modules/feed-ui/react/Representatives.tsx
@@ -1,10 +1,10 @@
-import { useState, type FC } from "react";
 import {
   LocationFilter,
   getLocationInformationText,
-} from "~app/modules/data/filters";
-import { getLegislators } from "~app/modules/data/representatives";
-import type { OfficialOffice } from "~app/modules/data/representatives/representatives.types";
+} from "@windycivi/domain/filters";
+import { getLegislators } from "@windycivi/domain/representatives";
+import type { OfficialOffice } from "@windycivi/domain/representatives/representatives.types";
+import { useState, type FC } from "react";
 import { DataField, ResultCard } from "~app/modules/design-system";
 import { FeedFilterProps } from "../feed-ui.types";
